error close + value checks for method arguments added
we now use errorClose() with a context to get the line/col # before
exiting. Still need to write the cleanup code to ensure the Stateflow
model is unchanged when a script fails. Added failure conditions for
each method argument that are checked dynamically, as well as some
size checks. Need to go through those in more detail to ensure
they're all dealt with.