--fromMATLAB to get around bug. runHATL now calls properly
If MATLAB calls the HATL interpreter from a shell command and the HATL
interpreter attempts to connect to the original shared MATLAB
instance, it hangs. I think this is due to a deadlock between MATLAB
being busy and waiting for the script to complete, while the script
waits for the MATLAB engine to return.

Need to run "initReaffirm" in the REAFFIRM_ROOT directory to have
things work. HATL expects REAFFIRM_ROOT to be set.